refactor(rig-1176): unify prompt hook interfaces
