+++ rig/rig-core/src/agent/mod.rs
@@ -114,9 +114,9 @@ mod tool;
 pub use crate::message::Text;
 pub use builder::{AgentBuilder, AgentBuilderSimple};
 pub use completion::Agent;
+pub use prompt_request::hooks::{HookAction, PromptHook, ToolCallHookAction};
 pub use prompt_request::streaming::{
     FinalResponse, MultiTurnStreamItem, StreamingError, StreamingPromptRequest, StreamingResult,
     stream_to_stdout,
 };
-pub use prompt_request::{HookAction, PromptHook, StreamingPromptHook, ToolCallHookAction};
 pub use prompt_request::{PromptRequest, PromptResponse};
